collapse AbstractProject into ProjectImpl
